Fix migrations by deleting bad ONION architecture
