--- YNote.Domain/Constants.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace YNote.Domain
-{
-    public class Constants
-    {
-    }
-}
