--- YNote.Domain/Entities/Base/BaseEntity.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-namespace YNote.Domain.Entities.Base {
-    public abstract class BaseEntity {
-        public int Id { get; set; }
-    }
-}
