--- YNote.Domain/Entities/TaskEntity.cs
@@ -1,15 +1,0 @@
-﻿using YNote.Domain.Entities.Base;
-
-namespace YNote.Domain.Entities {
-    public class TaskEntity:BaseEntity {
-
-        public int NoteId { get; set; }
-        public NoteEntity Note { get; set; }
-
-        public string SumUp { get; set; }
-
-        public string Description { get; set; }
-
-        public bool Status { get; set; }
-    }
-}
