--- YNote.Domain/Entities/UserEntity.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using YNote.Domain.Entities.Base;
-
-namespace YNote.Domain.Entities {
-    public class UserEntity{
-
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public string Surname { get; set; }
-        public string Nickname { get; set; }
-        public string Email { get; set; }
-        public string Password { get; set; }
-
-        public List<NoteEntity> Notes { get; set; }
-        public List<SpaceEntity> Spaces { get; set; }
-    }
-}
