--- YNote.Mappers/YNote.Mappers.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\YNote.Domain\YNote.Domain.csproj" />
-    <ProjectReference Include="..\YNote.Services.Models\YNote.Services.Models.csproj" />
-    <ProjectReference Include="..\YNote.WebAPI.Models\YNote.WebAPI.Models.csproj" />
-  </ItemGroup>
-
-</Project>
