--- YNote.Services.Interfaces/YNote.Services.Interfaces.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\YNote.Services.Models\YNote.Services.Models.csproj" />
-  </ItemGroup>
-
-</Project>
