--- YNote.Services/YNote.Services.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\YNote.Domain\YNote.Domain.csproj" />
-    <ProjectReference Include="..\YNote.Infrastructure\YNote.Infrastructure.csproj" />
-    <ProjectReference Include="..\YNote.Services.Interfaces\YNote.Services.Interfaces.csproj" />
-    <ProjectReference Include="..\YNote.Services.Models\YNote.Services.Models.csproj" />
-  </ItemGroup>
-
-</Project>
