--- YNote.WebAPI/YNote.WebAPI.csproj
@@ -1,28 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.2.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Areas\YNote.Domain\Controllers\" />
-    <Folder Include="Areas\YNote.Domain\Data\" />
-    <Folder Include="Areas\YNote.Domain\Models\" />
-    <Folder Include="Areas\YNote.Domain\Views\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\YNote.IoC\YNote.IoC.csproj" />
-    <ProjectReference Include="..\YNote.Services.Interfaces\YNote.Services.Interfaces.csproj" />
-    <ProjectReference Include="..\YNote.Services.Models\YNote.Services.Models.csproj" />
-    <ProjectReference Include="..\YNote.WebAPI.Models\YNote.WebAPI.Models.csproj" />
-  </ItemGroup>
-
-</Project>
