--- YNote.WebAPI/appsettings.json
@@ -1,13 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Warning"
-    }
-  },
-  
-  "ConnectionStrings": {
-    "DefaultConnection": "Data Source=DESKTOP-8DEVSOF;Initial Catalog=YNote;Integrated Security=True;"
-  }
-  ,
-  "AllowedHosts": "*"
-}
