+++ YNote.sln
@@ -3,23 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.779
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.Domain", "YNote.Domain\YNote.Domain.csproj", "{82D960BD-3C79-4B9B-A279-8B94594EF314}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.Infrastructure", "YNote.Infrastructure\YNote.Infrastructure.csproj", "{AC165B70-0231-409F-BDC6-DF31B3F14584}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.IoC", "YNote.IoC\YNote.IoC.csproj", "{E9EAEE13-33B2-4F67-BCF1-66F0DB5DC7F9}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.Mappers", "YNote.Mappers\YNote.Mappers.csproj", "{90F9C10B-E845-42C6-99F2-849FD9280457}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.Services", "YNote.Services\YNote.Services.csproj", "{259FA0FA-8270-418F-803C-7F6FB05F256F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.Services.Interfaces", "YNote.Services.Interfaces\YNote.Services.Interfaces.csproj", "{3F42115A-8DD8-40E3-83AD-D9E58EFFDA7F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.Services.Models", "YNote.Services.Models\YNote.Services.Models.csproj", "{4912A01D-F447-40C1-ACDD-E1D5912C0C54}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.WebAPI.Models", "YNote.WebAPI.Models\YNote.WebAPI.Models.csproj", "{34A4D86C-D12C-4078-B438-CE9E8B3DF876}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YNote.WebAPI", "YNote.WebAPI\YNote.WebAPI.csproj", "{A23C965B-94CC-47FC-A9C9-28601FE16359}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "YNote", "YNote\YNote.csproj", "{EAD9DE98-2CDA-4CF4-8D58-C684D1FE7904}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,47 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{82D960BD-3C79-4B9B-A279-8B94594EF314}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{82D960BD-3C79-4B9B-A279-8B94594EF314}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{82D960BD-3C79-4B9B-A279-8B94594EF314}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{82D960BD-3C79-4B9B-A279-8B94594EF314}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AC165B70-0231-409F-BDC6-DF31B3F14584}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AC165B70-0231-409F-BDC6-DF31B3F14584}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AC165B70-0231-409F-BDC6-DF31B3F14584}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AC165B70-0231-409F-BDC6-DF31B3F14584}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E9EAEE13-33B2-4F67-BCF1-66F0DB5DC7F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E9EAEE13-33B2-4F67-BCF1-66F0DB5DC7F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E9EAEE13-33B2-4F67-BCF1-66F0DB5DC7F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E9EAEE13-33B2-4F67-BCF1-66F0DB5DC7F9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{90F9C10B-E845-42C6-99F2-849FD9280457}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{90F9C10B-E845-42C6-99F2-849FD9280457}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{90F9C10B-E845-42C6-99F2-849FD9280457}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{90F9C10B-E845-42C6-99F2-849FD9280457}.Release|Any CPU.Build.0 = Release|Any CPU
-		{259FA0FA-8270-418F-803C-7F6FB05F256F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{259FA0FA-8270-418F-803C-7F6FB05F256F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{259FA0FA-8270-418F-803C-7F6FB05F256F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{259FA0FA-8270-418F-803C-7F6FB05F256F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3F42115A-8DD8-40E3-83AD-D9E58EFFDA7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F42115A-8DD8-40E3-83AD-D9E58EFFDA7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F42115A-8DD8-40E3-83AD-D9E58EFFDA7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F42115A-8DD8-40E3-83AD-D9E58EFFDA7F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4912A01D-F447-40C1-ACDD-E1D5912C0C54}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4912A01D-F447-40C1-ACDD-E1D5912C0C54}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4912A01D-F447-40C1-ACDD-E1D5912C0C54}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4912A01D-F447-40C1-ACDD-E1D5912C0C54}.Release|Any CPU.Build.0 = Release|Any CPU
-		{34A4D86C-D12C-4078-B438-CE9E8B3DF876}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{34A4D86C-D12C-4078-B438-CE9E8B3DF876}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{34A4D86C-D12C-4078-B438-CE9E8B3DF876}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{34A4D86C-D12C-4078-B438-CE9E8B3DF876}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A23C965B-94CC-47FC-A9C9-28601FE16359}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A23C965B-94CC-47FC-A9C9-28601FE16359}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A23C965B-94CC-47FC-A9C9-28601FE16359}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A23C965B-94CC-47FC-A9C9-28601FE16359}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EAD9DE98-2CDA-4CF4-8D58-C684D1FE7904}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EAD9DE98-2CDA-4CF4-8D58-C684D1FE7904}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EAD9DE98-2CDA-4CF4-8D58-C684D1FE7904}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EAD9DE98-2CDA-4CF4-8D58-C684D1FE7904}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7464261B-8DBF-4917-A194-B6D47AD0259D}
+		SolutionGuid = {9E1BB9B9-43C3-4D63-BC8A-6252D47E2E12}
 	EndGlobalSection
 EndGlobal
