+++ YNote/Entities/Base/BaseEntity.cs
@@ -1,0 +1,5 @@
+﻿namespace YNote.Entities.Base {
+    public abstract class BaseEntity {
+        public int Id { get; set; }
+    }
+}
