+++ YNote/Entities/TaskEntity.cs
@@ -1,0 +1,15 @@
+﻿using YNote.Entities.Base;
+
+namespace YNote.Entities {
+    public class TaskEntity:BaseEntity {
+
+        public int NoteId { get; set; }
+        public NoteEntity Note { get; set; }
+
+        public string SumUp { get; set; }
+
+        public string Description { get; set; }
+
+        public bool Status { get; set; }
+    }
+}
