+++ YNote/appsettings.Development.json
